L1v4 numberOfPoints vs. numberOfSteps change.
This change unifies the library so that it only stores 'numberOfSteps' in one place, reads either numberOfPoints or numberOfSteps, and writes numberOfPoints or numberOfSteps, depending on the level/version of the document.

This also adds this change to the 'UniformRange' object, where it was missed the first time through the spec.