Move all id/name functions to SedBase.cpp
This change moves all functionality with respect to id and name to SedBase, so the bulk of the change is simply 'deleted code'.

In L1v4, everything has an id and name, so everything works fine there.

In L1v3 and earlier, some elements had id and name, some had just id, some had just name, and some had neither.  This is now handled with the 'mIdAllowedPreV4' and 'mNameAllowedPreV4' boolean member variables on SedBase, which the relevant classes define in their constructors, and then checked in the various SedBase functions that set/get/unset/read/write the id and name attributes.  The exception is that the 'id is required' functionality is still handled in 'readAttributes', but instead of setting the id and then checking if it's set, it now just calls SedBase::readAttributes and then checks 'isSetId()'.