Remove the 'dist' target.
'dist' is a holdover from gnumake days, and is no longer used in modern CMake files.  It actually causes a problem, since when I add both this and libSEDML and/or libSBML to a 'x-deps' project, the 'dist' targets conflict with one another.  I think it's safest to just drop it, though it could be renamed if needed.  If it was super crucial, you could check to make sure no other 'dist' didn't already exist, and add it if so?  But I think that's overkill.